chore: #161 move source files to /src folder
