+++ .github/workflows/testing.yaml
@@ -20,10 +20,13 @@ jobs:
           go-version: '1.23.2'
 
       - name: Install dependencies
+        working-directory: ./src
         run: go get .
 
       - name: Build
+        working-directory: ./src
         run: go build
 
       - name: Test with the Go CLI
+        working-directory: ./src
         run: go test
