+++ src/errorPages/config.go
