+++ src/errorPages/errorPage.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/sevensolutions/traefik-oidc-auth/logging"
+	"github.com/sevensolutions/traefik-oidc-auth/src/logging"
 )
 
 type ProblemDetails struct {
