+++ src/predicate/lib.go
