+++ src/predicate/parse.go
