+++ src/predicate/predicate.go
